FEATURE: Add Texture and switch timing to double; fix camera call
Add a new Texture header (mapi/Texture.h) providing a basic Texture class and register it in the Visual Studio project file. Change time-related types from float to double (mapi/System.h) and remove the float cast in System.cpp to use glfwGetTime() with higher precision. Update World API to accept double deltaTime (mapi/World.h / mapi/World.cpp) and replace cam->computeModelMatrix() with cam->computeProjectionMatrix() during camera updates. These changes improve time precision and add the initial texture abstraction.

+++ mapi/System.cpp
@@ -85,7 +85,7 @@ void System::mainLoop()
 
 	while (!end)
 	{
-		newTime = static_cast<float>(glfwGetTime());
+		newTime = glfwGetTime();
 		deltaTime = newTime - lastTime;
 		lastTime = newTime;
